Moved files into packages; implemented specs2 tests
