+++ play-interop-test-scala/src/test/scala/play/api/test/NewTestServer.scala
@@ -1,0 +1,11 @@
+/*
+ * Copyright (C) 2018 Lightbend Inc. <https://www.lightbend.com>
+ */
+
+package play.api.test
+
+// TODO: Merge this functionality into TestServer itself
+final case class NewTestServer(
+  testServer: TestServer,
+  endpoints: ServerEndpoints,
+  stopServer: AutoCloseable)
